+++ client/src/canvas2d.cc
@@ -47,12 +47,9 @@ void internal::DataSetBuffer<TemporalCanvas2DComponent>::onSync() {
         cb->data_ids.push_back(msg_data->id);
     }
     data->canvas2d_store.setSend(target_, cb);
-    std::shared_ptr<std::function<void(Canvas2D)>> change_event;
-    {
-        std::lock_guard lock(data->event_m);
-        change_event =
-            data->canvas2d_change_event[target_.member_][target_.field_];
-    }
+    auto change_event =
+        internal::findFromMap2(data->canvas2d_change_event.shared_lock().get(),
+                               target_.member_, target_.field_);
     if (change_event && *change_event) {
         change_event->operator()(target_);
     }
@@ -61,8 +58,7 @@ const Canvas2D &
 Canvas2D::onChange(std::function<void(Canvas2D)> callback) const {
     this->request();
     auto data = dataLock();
-    std::lock_guard lock(data->event_m);
-    data->canvas2d_change_event[this->member_][this->field_] =
+    data->canvas2d_change_event.lock().get()[this->member_][this->field_] =
         std::make_shared<std::function<void(Canvas2D)>>(std::move(callback));
     return *this;
 }