+++ common/include/webcface/common/encoding.h
@@ -1,6 +1,4 @@
 #pragma once
-#include <map>
-#include <set>
 #include <string>
 #include <string_view>
 #include <memory>
@@ -109,13 +107,6 @@ class WEBCFACE_DLL SharedString {
     };
 };
 
-template <typename T>
-using StrMap1 = std::map<SharedString, T>;
-template <typename T>
-using StrMap2 = StrMap1<StrMap1<T>>;
-using StrSet1 = std::set<SharedString>;
-using StrSet2 = StrMap1<StrSet1>;
-
 namespace [[deprecated("symbols in webcface::encoding namespace are "
                        "now directly in webcface namespace")]] encoding {
 inline bool usingUTF8() { return webcface::usingUTF8(); }
@@ -128,13 +119,6 @@ inline std::string toNarrow(std::wstring_view name) {
 }
 
 using SharedString = webcface::SharedString;
-
-template <typename T>
-using StrMap1 = webcface::StrMap1<T>;
-template <typename T>
-using StrMap2 = webcface::StrMap2<T>;
-using StrSet1 = webcface::StrSet1;
-using StrSet2 = webcface::StrSet2;
 
 } // namespace encoding