+++ common/src/fmt.cc
@@ -66,6 +66,10 @@ WEBCFACE_MESSAGE_FMT_DEF(webcface::message::Sync) {
     return fmt::format_to(ctx.out(), "{}-Sync(member_id={}, time={:%T})",
                           msg_kind, m.member_id, m.getTime());
 }
+WEBCFACE_MESSAGE_FMT_DEF(webcface::message::Closed) {
+    return fmt::format_to(ctx.out(), "{}-Closed(member_id={})", msg_kind,
+                          m.member_id);
+}
 
 /// \private
 template <typename T,
@@ -290,8 +294,9 @@ WEBCFACE_MESSAGE_FMT_DEF_REQ(Canvas2DOld)
 
 WEBCFACE_MESSAGE_FMT_DEF(webcface::message::FuncInfo) {
     return fmt::format_to(
-        ctx.out(), "{}-FuncInfo('{}' from member_id={}, {} args, return={})",
-        msg_kind, m.field.decode(), m.member_id, m.args.size(),
+        ctx.out(),
+        "{}-FuncInfo('{}' from member_id={}, index={}, {} args, return={})",
+        msg_kind, m.field.decode(), m.member_id, m.index, m.args.size(),
         webcface::valTypeStr(m.return_type));
 }
 WEBCFACE_MESSAGE_FMT_DEF(webcface::message::Call) {