+++ CHANGELOG.md
@@ -2,16 +2,26 @@
 <!-- * リリースにおいて重要度の低い変更はコメントアウトで記述 -->
 <!-- * ドキュメントやREADMEの修正等は記述しない -->
 
-## [2.8.0] -
+## [2.9.0] - 2025-02-09
+### Added
+* Closedメッセージ (#503)
+	* Member::exists(), connected(), onConnect(), onDisconnect()
+	* Client::onConnect(), onDisconnect()
+<!-- * rust風のmutex管理クラス internal::MutexProxy, internal::SharedMutexProxy をつくった -->
+### Fixed
+* マルチスレッドでのWebCFaceクライアントの初期化処理を修正 (#504)
+
+## [2.8.0] - 2025-01-30
 ### Changed
 * webcface-common のビルドにも `-fabi-version=11`と`-Wabi`の指定を追加 (#476)
 <!-- * Ubuntu arm64とarmhfのテストをするCIを追加 (#478) -->
-* macOS以外でも libcurl >=8.10 を除外する制約を追加 (#484)
 <!-- ### Fixed -->
 ### Added
 * installer.sh (#480, #481)
 * ValueElementRef, Value::at(), size() (#486)
 	* `Value::operator[](int)` の返り値型が変更になる
+* FuncをWebUI上で名前順ではなく登録順に並べる機能を追加 (#487)
+	* Func::index(), setIndex()
 
 ## [2.7.0] - 2025-01-17
 ### Changed